requirements.txt
alembic==1.9.2
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.0
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
Pillow==9.4.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
